+++ packages/components-angular/src/app/pages/index.ts
@@ -22,16 +22,18 @@ export const pages = [
 export * from './core-initializer-prefixed.component';
 export * from './events.component';
 export * from './form-wrapper-binding.component';
-export * from './optional-properties.component';
-export * from './theme-injection.component';
-export * from './utilities.component';
-export * from './tabs-bar-navigation';
-export * from './tabs-bar-navigation-bug';
 
 /* Auto Generated Below */
 export * from './generated/core-initializer.component';
 export * from './generated/overview.component';
+export * from './optional-properties.component';
+export * from './tabs-bar-navigation';
+export * from './tabs-bar-navigation-bug';
+export * from './theme-injection.component';
+export * from './utilities.component';
+
 import { AccordionComponent } from './generated/accordion.component';
+import { AiTagComponent } from './generated/ai-tag.component';
 import { BannerComponent } from './generated/banner.component';
 import { BannerBasicComponent } from './generated/banner-basic.component';
 import { BannerErrorComponent } from './generated/banner-error.component';
@@ -244,10 +246,10 @@ import { TabsBarComponent } from './generated/tabs-bar.component';
 import { TagComponent } from './generated/tag.component';
 import { TagDismissibleComponent } from './generated/tag-dismissible.component';
 import { TextComponent } from './generated/text.component';
-import { TextareaComponent } from './generated/textarea.component';
-import { TextareaWrapperComponent } from './generated/textarea-wrapper.component';
 import { TextFieldWrapperComponent } from './generated/text-field-wrapper.component';
 import { TextListComponent } from './generated/text-list.component';
+import { TextareaComponent } from './generated/textarea.component';
+import { TextareaWrapperComponent } from './generated/textarea-wrapper.component';
 import { ToastComponent } from './generated/toast.component';
 import { ToastBasicComponent } from './generated/toast-basic.component';
 import { ToastMultilineComponent } from './generated/toast-multiline.component';
@@ -265,6 +267,7 @@ import { WordmarkComponent } from './generated/wordmark.component';
 
 export const generatedPages = [
   AccordionComponent,
+  AiTagComponent,
   BannerComponent,
   BannerBasicComponent,
   BannerErrorComponent,
@@ -502,6 +505,11 @@ export const generatedRoutes: ExtendedRoute[] = [
     name: 'Accordion',
     path: 'accordion',
     component: AccordionComponent,
+  },
+  {
+    name: 'Ai Tag',
+    path: 'ai-tag',
+    component: AiTagComponent,
   },
   {
     name: 'Banner',