+++ packages/components-js/projects/jsdom-polyfill/tests/unit/helper/index.ts
@@ -204,6 +204,7 @@ export const tagNameMarkup: Record<TagNameRelevant, string> = {
   </p-textarea-wrapper>`,
   'p-toast': '<p-toast></p-toast>',
   'p-wordmark': '<p-wordmark></p-wordmark>',
+  'p-ai-tag': '<p-ai-tag></p-ai-tag>',
 };
 
 export const getMarkup = (tagName: TagName): string => tagNameMarkup[tagName];